% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odes.R
\name{odes}
\alias{odes}
\title{ODE}
\usage{
odes(y, times, func, parms, var, range_var, ...)
}
\arguments{
\item{y}{}

\item{times}{}

\item{func}{}

\item{parms}{}

\item{var}{}

\item{range_var}{}

\item{...}{}
}
\description{
ODE
}
