% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_Hi}
\alias{parameters_Hi}
\title{Dose, ng, Prueksaritanont et al. (2014). Br.J.Clin.Pharmacol 78(3): 587-598}
\format{Parameters
\describe{
  \item{time}{Actual survival months observed}
  \item{cens}{censoring indicator 1 observed, 0 censored}
  \item{featurenames}{feature names in the gene expression matrix}
  ...
}}
\usage{
parameters_Hi
}
\description{
Dose, ng, Prueksaritanont et al. (2014). Br.J.Clin.Pharmacol 78(3): 587-598
}
\keyword{datasets}
